Refactor HeightMapGenerator using partial classes
