+++ CentrED/UI/Windows/HeightMapGenerator.Group.cs
@@ -1,0 +1,27 @@
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Threading.Tasks;
+using System.Text.Json;
+using CentrED.Client;
+using CentrED.Client.Map;
+using CentrED.IO.Models;
+using CentrED.Network;
+using ImGuiNET;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+using static CentrED.Application;
+
+namespace CentrED.UI.Windows;
+
+public partial class HeightMapGenerator
+{
+    private class Group
+    {
+        public float Chance = 100f;
+        public sbyte MinHeight = -128;
+        public sbyte MaxHeight = 127;
+        public List<ushort> Ids = new();
+    }
+}
